+++ lib/floating.dart
@@ -7,3 +7,4 @@ import 'package:flutter/widgets.dart';
 
 part 'src/floating.dart';
 part 'src/pip_switcher.dart';
+part 'src/enable_arguments.dart';